feat(projects): move oss project to `/projects/oss)
